Fix boot reset for object outlining state
